requirements.txt
pyannote.audio
pyannote_whisper
pywhispercpp
os
pydub